backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
